requirements.txt
python-telegram-bot==20.7
openai==1.3.7
google-generativeai>=0.3.0
google-auth>=2.15.0
google-cloud-vision>=3.4.0
requests>=2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
beautifulsoup4==4.12.2
numpy>=1.24.0
psutil>=5.9.0
httpx>=0.25.2 
paho-mqtt>=1.6.1
opencv-python>=4.8.0
gymnasium==0.29.1
pytest-asyncio>=0.23.6
hypothesis>=6.100.0
msgspec>=0.18.6
PyYAML>=6.0.2